CCBC-213: Unify connection routines
The basic idea is to unify the structures for the connections used for
all our objects which depend on their own TCP connections. This is a
work in progress, though it does pass all unit tests without issue
(though integration tests are needed). Constructive comments
appreciated.

The history here was largely borrowed from another rewrite

Change-Id: Iffa9e7b0d8787ae42731f2716d9dc4e9f8f8a314