Migrate to Mock 0.6
The new mock uses a synchronous request/response protocol over
line-delimited JSON objects.

Included is jsoncpp which should be pending approval.

The mock includes functionality for replica read as well as observe,
though this is not present in our unit tests (the actual tests for using
those features will be implemented later on).

Change-Id: Ibe41ed3955d44daf7f675dda92a3813bd81f5c18